refactor: move loading and saving teams to repo module
